% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{calc.chisq}
\alias{calc.chisq}
\title{Calculate chi-square value of an array or matrix}
\usage{
calc.chisq(A)
}
\arguments{
\item{A}{an array or matrix}
}
\value{
The chi-square value of the table
}
\description{
Calculate chi-square value of an array or matrix
}
\examples{
r <- c(2,3,4)
A <- array(1:prod(r),r)
calc.chisq(A)
}

