% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc.marg}
\alias{calc.marg}
\title{Calculate marginal counts of array}
\usage{
calc.marg(A)
}
\arguments{
\item{A}{An array or matrix}
}
\value{
A list of marginal counts per variable
}
\description{
Calculate marginal counts of array
}
\examples{
r <- c(2,3,4)
A <- array(1:prod(r),r)
calc.marg(A)
}

