% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc.pr}
\alias{calc.pr}
\title{Calculate probability indicated by segments in random unit vector directions}
\usage{
calc.pr(segs.out, n, df, one.side)
}
\arguments{
\item{segs.out}{Output of window.select}

\item{n}{NUmber of random unit vectors}

\item{df}{Degrees of freedom}

\item{one.side}{Optional. Logical. When TRUE, test vectors indicate one-side test and otherwise two-sided}
}
\value{
Probability estimated
}
\description{
Calculate probability indicated by segments in random unit vector directions
}
\examples{
# utility function being used in other functions
}

