% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{CategoryVector}
\alias{CategoryVector}
\title{Coordinate of simplex vertices}
\usage{
CategoryVector(d)
}
\arguments{
\item{d}{Number of vertices of a simplex}
}
\value{
d times (d-1) matrix, each row of which is a coordinate of a vertex of (d-1)-simplex, whose center is the origin and distance of whose vertices from the origin is 1
}
\description{
Coordinate of simplex vertices
}
\examples{
d <- 3
cv <- CategoryVector(d)
print(cv)
apply(cv^2,1,sum)
cv \%*\% t(cv) 
}

