% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{make.alternative.table}
\alias{make.alternative.table}
\title{Generate a table indicating alternative hypothesis}
\usage{
make.alternative.table(t.sphere, test.table, odds.ratio.table, or, n, K)
}
\arguments{
\item{t.sphere}{Output of table.sphere of a table of interest itself or a table sharing marginal counts with the table of interest}

\item{test.table}{A matrix or array of a table which indicates the direction of alternative hypothesis along with an argument odds.ratio.table}

\item{odds.ratio.table}{A matrix or array of a table with cell values being 0,1,2,3 or 4: Odds ratio of table is indicated by the (sum(cells of 1)*sum(cells of 4))/(sum(cells of 2)*sum(cells of 3))}

\item{or}{A number indicating odds ratio defined by odds.ratio.table argument}

\item{n}{Optional. An integer which makes a one-dimensional grid with n points, among which the table with odds ratio closest to the argument or is selected}

\item{K}{Optional. A number indicating the maximum value of square of chi-square value to search the table of alternative hypothsis}
}
\value{
alt.table The estimated table of alternative hypothesis

alt.vec The vector of the estimated table in df-dimensional space
}
\description{
Generate a table indicating alternative hypothesis
}
\examples{
A <- matrix(c(10,20,30,40,50,60),2,3)
t.sphere <- table.sphere(A)
test.table <- matrix(c(1,0.5,0,0,0,0),byrow=TRUE,2,3)
odds.ratio.table <- matrix(c(1,0,2,3,0,4),byrow=TRUE,2,3)
or <- 2
n <- 1000
K <- 10
make.alternative.table(t.sphere,test.table,odds.ratio.table,or,n,k)
}

