% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.exp.table}
\alias{make.exp.table}
\title{Make an expected table}
\usage{
make.exp.table(A)
}
\arguments{
\item{A}{an array or matrix}
}
\value{
The expected table in a shape of array or matrix
}
\description{
Make an expected table
}
\examples{
r <- c(2,3,4)
A <- array(1:prod(r),r)
make.exp.table(A)
}

