% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.simplex.multi}
\alias{make.simplex.multi}
\title{Kronecker product of simplex-based rotation matrices}
\usage{
make.simplex.multi(r)
}
\arguments{
\item{r}{A vector of positive integers, a vector of number of levels of variables}
}
\value{
Kronecker product of rotation matrices, each of which is a simplex-based rotation matrix of number of levels of each variable
}
\description{
Kronecker product of simplex-based rotation matrices
}
\examples{
r <- c(2,3,4)
KM <- make.simplex.multi(r)
dim(KM)
}

