% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{make.table.from.data}
\alias{make.table.from.data}
\title{Generate an array or matrix from data matrix}
\usage{
make.table.from.data(d)
}
\arguments{
\item{d}{a matrix of sample size x number of variables, levels of which should be integers starting from 1}
}
\value{
An array or matrix
}
\description{
Generate an array or matrix from data matrix
}
\examples{
m <- matrix(c(10,20,30,40,50,60),2,3)
d <- make.data.from.table(m)
make.table.from.data(d)
}

