% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{make.vector0}
\alias{make.vector0}
\title{Make a vector indexing rows that should be zero}
\usage{
make.vector0(r)
}
\arguments{
\item{r}{A vector of positive integers, a vector of number of levels of variables}
}
\value{
A vector with 0 or 1; indices with 1 means that elements of table vectors should be zero after the kronecker product rotation.
}
\description{
Make a vector indexing rows that should be zero
}
\examples{
r <- c(2,3,4)
v0 <- make.vector0(r)
}

