% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.X.vector0}
\alias{make.X.vector0}
\title{Multiway table-related rotation matrix and its features}
\usage{
make.X.vector0(r)
}
\arguments{
\item{r}{A vector of positive integers, a vector of number of levels of variables}
}
\value{
X A Rotation matrix; output of make.simplex.multi()

X.sub A matrix only with meaningful rows of X

X.inv Inverse matrix of X

X.inv.sub Meaningful part of X.inv

vector0 A vector indicating meaningful rows of X
}
\description{
Multiway table-related rotation matrix and its features
}
\examples{
r <- c(2,3,4)
mXv <- make.X.vector0(r)
mXv$X \%*\% mXv$X.inv
}

