% Generated by roxygen2 (4.0.1): do not edit by hand
\name{map.matrix}
\alias{map.matrix}
\title{Make three matrices that are the output of map2df(), map2full() and map2dfdir}
\usage{
map.matrix(A)
}
\arguments{
\item{A}{An array or matrix of multiway table}
}
\value{
Full2DF A matrix transferring table vectors to df-dimension vectors

DF2full A matrix transferring back vectors in df-dimensional space back to vectors with number-of-cell elements

Normalvec2DF A matric transferring test vectors to df-dimension vectors
}
\description{
Make three matrices that are the output of map2df(), map2full() and map2dfdir
}
\examples{
r <- c(2,3,4)
A <- array(1:prod(r),r)
map.matrix(A)
}

