% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{map2dfdir}
\alias{map2dfdir}
\title{Make a matrix that transfer test vectors to the direction in the df-dimensional space}
\usage{
map2dfdir(A)
}
\arguments{
\item{A}{An array or matrix of multiway table}
}
\value{
A matrix
}
\description{
Make a matrix that transfer test vectors to the direction in the df-dimensional space
}
\examples{
r <- c(2,3,4)
A <- array(1:prod(r),r)
M <- map2dfdir(A)
c(A) \%*\% M
}

