% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{map2full}
\alias{map2full}
\title{Make a matrix that transfer back a vector in the df-dimensional space to a vector with number-of-cell of table elements}
\usage{
map2full(A)
}
\arguments{
\item{A}{An array or matrix of multiway table}
}
\value{
A matrix
}
\description{
Make a matrix that transfer back a vector in the df-dimensional space to a vector with number-of-cell of table elements
}
\examples{
r <- c(2,3,4)
A <- array(1:prod(r),r)
M <- map2df(A)
M.inv <- map2full(A)
M.inv \%*\% (M \%*\% c(A))
}

