% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{model.table.series}
\alias{model.table.series}
\title{Generate althernative hypothesis tables in a direction with multipls K values}
\usage{
model.table.series(t.sphere, tests, test.table, odds.ratio.table, ks)
}
\arguments{
\item{t.sphere}{Output of table.sphere}

\item{tests}{A list of tables indicating tests}

\item{test.table}{An array or matrix of table indicating the direction of alternaitve hypothesis}

\item{odds.ratio.table}{An array or matrix of table indicating 1,2,3,4 cells for calculation of odds ratio to be controled}

\item{ks}{Optional A vector of K statistic values}
}
\value{
\itemize{
\item{"alt.tables"}{A list of alternative tables}
\item{"alt.vecs"}{A matrix of alternative table vectors in df-dimensional space}
\item{"alt.intersect.out"}{A matrix of terms used to calculate pertinent information}
\item{"stats"}{A vector of max K of alternative tables}
\item{"difs"}{A matrix each row of which has difference of cells of alternative hypothesis tables}
\item{"ors"}{A vector of odds ratios}
}
}
\description{
Generate althernative hypothesis tables in a direction with multipls K values
}
\examples{
# Utility function for the authors

}

