% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mway.test}
\alias{mway.test}
\title{Max test}
\usage{
mway.test(A, tests, lower.tail, one.side, n)
}
\arguments{
\item{A}{An array or matrix of table that specifies marginal counts}

\item{tests}{A list of tables}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X ?????? x], otherwise, P[X > x].}

\item{one.side}{Optional. When TRUE, tests are one-sided, otherwise two-sided}

\item{n}{number of random unit vectors used for estimation}
}
\value{
\itemize{
\item{"statistic"}{maximum among test statistics; K^2-equivalent}
\item{"p.value"}{multiple testing-corrected P value}
}
}
\description{
Max test
}
\examples{
A <- matrix(c(10,20,30,40,50,60),2,3)
t1 <- t2 <- t3 <- matrix(c(1,0,0,0,0,0),byrow=TRUE,2,3)
t2[1,2] <- 0.5
t3[1,2] <- 1
tests <- list(t1,t2,t3)
n <- 1000
mway.test(A,tests,n=10000)
}

