% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{pmway.intersect}
\alias{pmway.intersect}
\title{Power calculator with intersect information rather than alternative hypothesis table given}
\usage{
pmway.intersect(x, rmway.out, intersect, lower.tail)
}
\arguments{
\item{x}{K statistic values}

\item{rmway.out}{Output of rmway function}

\item{intersect}{alternative hypothesis-dependent term}

\item{lower.tail}{ogical; if TRUE (default), probabilities are P[X ?????? x], otherwise, P[X > x].}
}
\value{
Power
}
\description{
Power calculator with intersect information rather than alternative hypothesis table given
}
\examples{
# Utility function for the authors
}

