% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{pmway}
\alias{pmway}
\title{Randomize a data matrix}
\usage{
pmway(x, rmway.out, alt.intersect, lower.tail)
}
\arguments{
\item{x}{K statistic value}

\item{rmway.out}{output of rmway function}

\item{alt.intersect}{logical, if false (default), null hypotheis is assumued, otherwise alternative hypothesis is assumed}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X ?????? x], otherwise, P[X > x].}
}
\value{
probability
}
\description{
Randomize a data matrix
}
\examples{
# Utility function
}

