% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{pmway.table}
\alias{pmway.table}
\title{Estimate p-values/cumulative probability of maximum statistics on multi-way table under unll or alternative hypothesis}
\usage{
pmway.table(x, A, tests, lower.tail, nc, one.side, n)
}
\arguments{
\item{x}{A vector of K^2 quantiles}

\item{A}{An array or matrix of table that specifies marginal counts}

\item{tests}{A list of tables}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X ?????? x], otherwise, P[X > x].}

\item{nc}{Optional. if FALSE (default), random tables are in null hypothesis and otherwise in alternative hypothesis with its ccenter being table A}

\item{one.side}{Optional. When TRUE, tests are one-sided, otherwise two-sided}

\item{n}{number of random unit vectors used for estimation}
}
\value{
\itemize{
\item{"p.null"}{cumulative probability under null hypothesis}
\item{"p.alt"}{cumulative probability under alternative hypothesis}
}
}
\description{
Estimate p-values/cumulative probability of maximum statistics on multi-way table under unll or alternative hypothesis
}
\details{
This function is one of three major functions in this package. Functions for a distribution are rxxxx,dxxxx, pxxxx and qxxxx
in general and this package provides three of them, rmway.table, pmway.table and qmway.table functions.
}
\examples{
A <- matrix(c(10,20,30,40,50,60),2,3)
t1 <- t2 <- t3 <- matrix(c(1,0,0,0,0,0),byrow=TRUE,2,3)
t2[1,2] <- 0.5
t3[1,2] <- 1
tests <- list(t1,t2,t3)
n <- 1000
x <- 0:10
pmway.table(x,A,tests)

}

