% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{power.mway.table}
\alias{power.mway.table}
\title{Power calculator}
\usage{
power.mway.table(A, tests, alpha, n)
}
\arguments{
\item{A}{An array or matrix of a table representing alternative hypothesis}

\item{tests}{A list of array/marix of tables indicating tests}

\item{alpha}{type 1 error threshold}

\item{n}{Number of iteration to estimate}
}
\value{
\itemize{
 \item{"pvalues"}{null.p: estimated cumulative probability for null hypothesis when statistic K being x or higher 
 that is estimated for quantile of 1-alpha. alt.p: estimated value for alternative hypotheis }
 \item{"chisqs"}{Estimated K^2 value as a quantile of 1-alpha}
 \item{"alpha.given"}{Argument alpha}
 \item{"alpha.used"}{Estimated p.value for the estimated K^2 value}
 \item{"power"}{Power}
}
}
\description{
Power calculator
}
\examples{
A <- matrix(c(10,20,30,40,50,60),2,3)
t1 <- t2 <- t3 <- matrix(c(1,0,0,0,0,0),byrow=TRUE,2,3)
t2[1,2] <- 0.5
t3[1,2] <- 1
tests <- list(t1,t2,t3)
alpha <- 0.05
n <- 1000
power.mway.table(A,tests,alpha,n)
}

