% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{qmway.table.pre}
\alias{qmway.table.pre}
\title{Grossly estimate quantile value. Utility function for qmway.table function}
\usage{
qmway.table.pre(p, A, tests, lower.tail, x, nc, one.side, n)
}
\arguments{
\item{p}{probability}

\item{A}{An array or matrix of table that specifies marginal counts}

\item{tests}{A list of tables}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X ?????? x], otherwise, P[X > x].}

\item{x}{a vector of K statistic values}

\item{nc}{if FALSE (default) null hypothesis, otherwise alternative hypothesis}

\item{one.side}{Optional. When TRUE, tests are one-sided, otherwise two-sided}

\item{n}{number of random unit vectors used for estimation}
}
\value{
two quantile values and two corresponding p values between which p and its quantile exist
}
\description{
Grossly estimate quantile value. Utility function for qmway.table function
}
\examples{
# Utility function
}

