% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reshape.table}
\alias{reshape.table}
\title{Utility function to handle tables with cells whose expected value is 0}
\usage{
reshape.table(A, A.ori)
}
\arguments{
\item{A}{an array or matrix of table, that may be the table with cells whose expected value is zero (0 cells) or
that may be the table without such cells}

\item{A.ori}{if missing(default), A represetnts a table with 0 cell(s) and the function converts A to a table without 0 cell(s).
Otherwise, A should be a table without 0 cell(s) and A.ori is the original array/matrix before shrinkage.
of table with 0 cells}
}
\value{
An array or matrix of table. if A.ori is missing, shrunken table is returned otherwise a table with 0 cells recovered.
with 0 cells.
}
\description{
Utility function to handle tables with cells whose expected value is 0
}
\examples{
n <- 5
M <- list()
N <-100
for(i in 1:n){
 M[[i]] <- sample(0:5,sample(2:4,1))
 M[[i]] <- M[[i]]/sum(M[[i]])*N
}
tmp <- M[[1]]
for(i in 2:length(M)){
  tmp <- t(M[[i]]/N) \%x\% tmp
}
A <- array(tmp,lapply(M,length))

A.r <- reshape.table(A)
A.2 <- reshape.table(A.r,A)
calc.marg(A)
calc.marg(A.r)
calc.marg(A.2)
A-A.2
}

