% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{rmway}
\alias{rmway}
\title{Generate random K statistic values}
\usage{
rmway(n, A, tests, alt.hx, odds.ratio.table, or, n.alt, K.alt)
}
\arguments{
\item{n}{Number of random unit vectors}

\item{A}{An array or matrix of a table}

\item{tests}{A list of test tables}

\item{alt.hx}{A matrix or array of a table which indicates the direction of alternative hypothesis along with an argument odds.ratio.table}

\item{odds.ratio.table}{A matrix or array of a table with cell values being 0,1,2,3 or 4: Odds ratio of table is indicated by the (sum(cells of 1)*sum(cells of 4))/(sum(cells of 2)*sum(cells of 3))}

\item{or}{A number indicating odds ratio defined by odds.ratio.table argument}

\item{n.alt}{Optional. An integer which makes a one-dimensional grid with n points, among which the table with odds ratio closest to the argument or is selected}

\item{K.alt}{Optional. A number indicating the maximum value of square of chi-square value to search the table of alternative hypothsis}
}
\value{
\itemize{
 \item{"t.sphere"}{ Output of table.sphere(A)}
 \item{"alt.table"}{Argument alt.table itself}
 \item{"tests"}{Argument tests}
 \item{"test.vecs"}{Matrix whose rows are test vectors}
 \item{"runitstat.out"}{Output of runitstat function}
 \item{"alt.intersect.out"}{Output of alt.intersect function}
 \item{"n"}{Argument n}
 \item{"A"}{Argument A}
 \item{"alt.hx"}{Argument alt.hx}
 \item{"odds.ratio.table"}{Argument odds.ratio.table}
 \item{"or"}{Argument or}
 \item{"n.alt"}{Argument n.alt}
 \item{"K.alt"}{Argument K.alt}
 \item{"rmvn"}{A matrix of random unit vectors, each row of which is a random unit vector}
 }
}
\description{
Generate random K statistic values
}
\examples{
A <- matrix(c(10,20,30,40,50,60),2,3)
t1 <- t2 <- t3 <- matrix(c(1,0,0,0,0,0),byrow=TRUE,2,3)
t2[1,2] <- 0.5
t3[1,2] <- 1
tests <- list(t1,t2,t3)
n <- 1000
rmway.out <- rmway(n,A,tests)
}

