% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{rmway.table}
\alias{rmway.table}
\title{Generate random tables under unll or alternative hypothesis}
\usage{
rmway.table(n, A, tests, nc, one.side)
}
\arguments{
\item{n}{Number of random tables to be generated}

\item{A}{An array or matrix of table that specifies marginal counts and also alternative hypothesis if nc is TRUE}

\item{tests}{Optional. When TRUE, test statistics and maximum among them are calculated and returned}

\item{nc}{Optional. When FALSE, random tables are in null hypothesis and when TRUE in alternative hypothesis with its ccenter being table A}

\item{one.side}{Optional. When TRUE, tests are one-sided, otherwise two-sided}
}
\value{
\itemize{
\item{"stats.all"}{A matrix of K statistics of random tables}
\item{"stats"}{A matrix of K^2 statistics of random tables}
\item{"rtables"}{A list of random tables}
\item{"ellipsoid.loc"}{A matrix of coordinates of random tables that are in df-dimensional space but the contours are ellipsoids}
}
}
\description{
Generate random tables under unll or alternative hypothesis
}
\details{
This function is one of three major functions in this package. Functions for a distribution are rxxxx,dxxxx, pxxxx and qxxxx
in general and this package provides three of them, rmway.table, pmway.table and qmway.table functions.
}
\examples{
A <- matrix(c(10,20,30,40,50,60),2,3)
n <- 1000
rmway.tables.null <- rmway.table(n,A)
rmway.tables.alt <- rmway.table(n,A,nc=TRUE)

}

