% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{runitstat}
\alias{runitstat}
\title{Generate random K statistic values for test vectors under null hypothesis}
\usage{
runitstat(test.vecs, rmvn, n)
}
\arguments{
\item{test.vecs}{A number-of-test rows x df columns matrix of test vectos, each row of which is a test vector}

\item{rmvn}{Optional A matrix of random unit vectors}

\item{n}{number of random unit vectors}
}
\value{
A number-of-unit-vector rows x number-of-test columns matrix of K statistic values
}
\description{
Generate random K statistic values for test vectors under null hypothesis
}
\examples{
A <- matrix(c(10,20,30,40,50,60),2,3)
t.sphere <- table.sphere(A)
t1 <- t2 <- t3 <- matrix(c(1,0,0,0,0,0),byrow=TRUE,2,3)
t2[1,2] <- 0.5
t3[1,2] <- 1
tests <- list(t1,t2,t3)
test.vecs <- make.test.vecs(t.sphere,tests)
runitstat(test.vecs)
}

