% Generated by roxygen2 (4.0.1): do not edit by hand
\name{segment.overlap}
\alias{segment.overlap}
\title{Overlapped segment(s) of multiple one-dimensional segments}
\usage{
segment.overlap(s)
}
\arguments{
\item{s}{A matrix with two columns each row of which indicating start and end of each segment}
}
\value{
A matrix with two columns each row of which indicating overlapped segment
}
\description{
Overlapped segment(s) of multiple one-dimensional segments
}
\examples{
s <- matrix(c(0,4,1,6,2,3),byrow=TRUE,ncol=2)
segment.overlap(s)
}

