% Generated by roxygen2 (4.0.1): do not edit by hand
\name{st.mvn}
\alias{st.mvn}
\title{Generate random unit vectors}
\usage{
st.mvn(n, df)
}
\arguments{
\item{n}{Number of random vecotrs to be generated}

\item{df}{An integer indicating dimension}
}
\value{
A n x df matrix, each row of which is a random unit vector
}
\description{
Generate random unit vectors
}
\examples{
n <- 100
df <- 2
plot(st.mvn(n,df))
}

