% Generated by roxygen2 (4.0.1): do not edit by hand
\name{window.ab}
\alias{window.ab}
\title{Generate terms to calculate K statistic values of tests for random unit vectors}
\usage{
\method{window}{ab}(as, ks, bs)
}
\arguments{
\item{as}{NUmber-of-random-unit-vector rows x number-of-test matrix}

\item{ks}{A vector or K values (square of chi-square equivalent values) for which P-values should be estimated}

\item{bs}{Optional. A Vector of K values of alternative hypothesis table for tests}
}
\value{
A list. k.per.a is 3-way array and is one set of terms to calculate K statistics, which is variable respect to unit vectors. b.per.a is another set of terms which K values for the alternative hypothesis itself
}
\description{
Generate terms to calculate K statistic values of tests for random unit vectors
}
\examples{
# utility function being used in other functions
}

