% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{window.select}
\alias{window.select}
\title{Calculate a segment to be integrated for individual K-cutoff value for each random unit vector}
\usage{
\method{window}{select}(window.ab.out, one.side)
}
\arguments{
\item{window.ab.out}{An output of window.ab}

\item{one.side}{Optional. Logical. When TRUE, test vectors indicate one-side test and otherwise two-sided}
}
\value{
A list segments is a matrix with two rows, 1st column of which is the start and 2nd is the end of segments; ids is a vector indicating random unit vector ids to which segments belong
}
\description{
Calculate a segment to be integrated for individual K-cutoff value for each random unit vector
}
\examples{
# utility function being used in other functions
}

