% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{calc.marg}
\alias{calc.marg}
\title{Calculate marginal counts of array}
\usage{
calc.marg(A)
}
\arguments{
\item{A}{An array or matrix}
}
\value{
A list of marginal counts per variable
}
\description{
Calculate marginal counts of array
}
\examples{
r <- c(2,3,4)
A <- array(1:prod(r),r)
calc.marg(A)
}

