% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.data.from.table}
\alias{make.data.from.table}
\title{Generate data matrix from table}
\usage{
make.data.from.table(m)
}
\arguments{
\item{m}{An array or matrix of table that specifies marginal counts}
}
\value{
A matrix of sample size x number of variables, levels of which are integers starting from 1
}
\description{
Generate data matrix from table
}
\examples{
m <- matrix(c(10,20,30,40,50,60),2,3)
make.data.from.table(m)
}

