% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{make.diff.table}
\alias{make.diff.table}
\title{Make a table, the given table minus its expected table}
\usage{
make.diff.table(A)
}
\arguments{
\item{A}{an array or matrix}
}
\value{
The difference table in a shape of array or matrix
}
\description{
Make a table, the given table minus its expected table
}
\examples{
r <- c(2,3,4)
A <- array(1:prod(r),r)
make.diff.table(A)
}

