% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{pchisq.bid}
\alias{pchisq.bid}
\title{Cumulative distribtution of bidirectional chisquare distribution supporting -Inf to +Inf}
\usage{
pchisq.bid(a, df)
}
\arguments{
\item{a}{Chi-square value or its negative}

\item{df}{Degrees of freedom}
}
\value{
Cumulative probability
}
\description{
Cumulative distribtution of bidirectional chisquare distribution supporting -Inf to +Inf
}
\examples{
as <- seq(from=-5,to=5,length=101)
plot(pchisq.bid(as,1))
}

