% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereKaji20140527.R
\name{segment.union}
\alias{segment.union}
\title{Union of segment(s) of multiple one-dimensional segments}
\usage{
segment.union(s)
}
\arguments{
\item{s}{A matrix with two columns each row of which indicating start and end of each segment}
}
\value{
A matrix with two columns each row of which indicating unified segment
}
\description{
Union of segment(s) of multiple one-dimensional segments
}
\examples{
s <- matrix(c(0,4,1,6,2,3),byrow=TRUE,ncol=2)
segment.union(s)
}

